+++ hyungjpa/utils/ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "../includes/test.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
